requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain
faiss-cpu
sentence-transformers
groq
